api/pyproject.toml
[tool.poetry]
name = "url shortener"
version = "0.1.0"
description = "URL Shortener project recruitment task."
authors = ["Martyna Witkowska <mar.witkowska@gmail.com>"]
license = "Beerware"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
Django = "4.2.1"
djangorestframework = "3.14.0"
pytest = "7.3.1"
pytest-django = "4.5.2"
pytest-cov = "4.0.0"
flake8 = "6.0.0"
flake8-django = "1.2"
black = {extras = ["d"], version = "^23.3.0"}
drf-spectacular = "0.26.2"
shortuuid = "1.0.11"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "config.settings"
python_files = ["test_*.py"]

[tool.black]
line-length = 120
target-version = ['py311']
exclude = '''
/(
    | \.git/  # exclude a directory
    | \.mypy_cache/  # exclude a directory
    | \.pytest_cache/  # exclude a directory
    | .*\.pyi?  # exclude generated files
    | .*\.json  # exclude metadata files
    | .*\.md  # exclude documentation
    | .*\.toml  # exclude configuration files
    | .*\.yml  # exclude configuration files
    | .*\.yaml  # exclude configuration files
    | .*\.txt  # exclude configuration files
    | .*\.lock  # exclude configuration files
    | .*\.sh  # exclude configuration files
    | .*\.log  # exclude configuration files
)/
'''

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
